Fixed problem with serial number assignation when creating a new database (not importing an old one)

+++ src/ca_file.h
@@ -74,7 +74,7 @@ gboolean ca_file_get_pkey_in_db_from_id (CaFileElementType type, guint64 db_id);
 gboolean ca_file_set_pkey_field_for_id (CaFileElementType type, const gchar *new_value, guint64 db_id);
 gboolean ca_file_mark_pkey_as_extracted_for_id (CaFileElementType type, const gchar *filename, guint64 db_id);
 
-gint ca_file_begin_new_crl_transaction (gint ca_id, time_t timestamp);
+gint ca_file_begin_new_crl_transaction (guint64 ca_id, time_t timestamp);
 void ca_file_commit_new_crl_transaction ();
 void ca_file_rollback_new_crl_transaction ();
